+++ lib/td_dd_web/views/data_structure_view.ex
@@ -69,7 +69,6 @@ defmodule TdDdWeb.DataStructureView do
       :deleted_at
     ])
     |> Map.put(:metadata, Map.get(data_structure, :metadata, %{}))
-    |> Map.put(:path, Map.get(data_structure, :path, []))
   end
 
   defp add_system_with_keys(json, data_structure, keys) do